+++ core/src/cloud/api.ts
@@ -9,8 +9,7 @@
 import { IncomingHttpHeaders } from "http"
 
 import { got, GotHeaders, GotHttpError, GotJsonOptions, GotResponse } from "../util/http"
-import { findProjectConfig } from "../config/base"
-import { CommandError, EnterpriseApiError } from "../exceptions"
+import { EnterpriseApiError } from "../exceptions"
 import { LogEntry } from "../logger/log-entry"
 import { gardenEnv } from "../constants"
 import type { ClientAuthToken as ClientAuthTokenType } from "../db/entities/client-auth-token"
@@ -18,12 +17,20 @@ import { Cookie } from "tough-cookie"
 import { isObject } from "lodash"
 import { deline } from "../util/string"
 import chalk from "chalk"
-import { GetProjectResponse, GetUserResponse } from "@garden-io/platform-api-types"
+import {
+  GetProjectResponse,
+  GetProfileResponse,
+  CreateProjectsForRepoResponse,
+  ListProjectsResponse,
+} from "@garden-io/platform-api-types"
 import { getCloudDistributionName, getPackageVersion } from "../util/util"
 import { CommandInfo } from "../plugin-context"
+import { ProjectResource } from "../config/project"
 
 const gardenClientName = "garden-core"
 const gardenClientVersion = getPackageVersion()
+
+export class EnterpriseApiDuplicateProjectsError extends EnterpriseApiError {}
 
 // If a GARDEN_AUTH_TOKEN is present and Garden is NOT running from a workflow runner pod,
 // switch to ci-token authentication method.
@@ -94,25 +101,57 @@ export interface RegisterSessionResponse {
   namespaceId: number
 }
 
+// Represents a cloud environment
+export interface CloudEnvironment {
+  id: number
+  name: string
+}
+
+// Represents a cloud project
+export interface CloudProject {
+  id: number
+  uid: string
+  name: string
+  repositoryUrl: string
+  environments: CloudEnvironment[]
+}
+
+function toCloudProject(
+  project: GetProjectResponse["data"] | ListProjectsResponse["data"][0] | CreateProjectsForRepoResponse["data"][0]
+): CloudProject {
+  const environments: CloudEnvironment[] = []
+
+  for (const environment of project.environments) {
+    environments.push({ id: environment.id, name: environment.name })
+  }
+
+  return {
+    id: project.id,
+    uid: project.uid,
+    name: project.name,
+    repositoryUrl: project.repositoryUrl,
+    environments,
+  }
+}
+
 /**
- * A helper function that finds a project without resolving template strings and returns the enterprise
- * config. Needed since the EnterpriseApi is generally used before initializing the Garden class.
+ * A helper function to get the cloud domain from a project config. Uses the env var
+ * GARDEN_CLOUD_DOMAIN to override a configured domain.
  */
-export async function getEnterpriseConfig(currentDirectory: string) {
-  const projectConfig = await findProjectConfig(currentDirectory)
+export function getGardenCloudDomain(projectConfig?: ProjectResource): string | undefined {
   if (!projectConfig) {
-    throw new CommandError(`Not a project directory (or any of the parent directories): ${currentDirectory}`, {
-      currentDirectory,
-    })
+    return undefined
   }
 
-  const projectId = projectConfig.id
-  if (!projectId || !projectConfig.domain) {
-    return
+  let cloudDomain: string | undefined
+
+  if (gardenEnv.GARDEN_CLOUD_DOMAIN) {
+    cloudDomain = new URL(gardenEnv.GARDEN_CLOUD_DOMAIN).origin
+  } else if (projectConfig.domain) {
+    cloudDomain = new URL(projectConfig.domain).origin
   }
 
-  const domain = new URL(projectConfig.domain).origin
-  return { domain, projectId }
+  return cloudDomain
 }
 
 /**
@@ -126,21 +165,20 @@ export class CloudApi {
   private log: LogEntry
   private intervalMsec = 4500 // Refresh interval in ms, it needs to be less than refreshThreshold/2
   private apiPrefix = "api"
-  private _project?: GetProjectResponse["data"]
-  private _profile?: GetUserResponse["data"]
+  private _project?: CloudProject
+  private _profile?: GetProfileResponse["data"]
   public domain: string
-  public projectId: string
+  public projectId: string | undefined
 
   // Set when/if the Core session is registered with Cloud
   public environmentId?: number
   public namespaceId?: number
   public sessionRegistered = false
 
-  constructor(log: LogEntry, enterpriseDomain: string, projectId: string) {
+  constructor(log: LogEntry, enterpriseDomain: string) {
     this.log = log
     // TODO: Replace all instances of "enterpriseDomain" with "cloudDomain".
     this.domain = enterpriseDomain
-    this.projectId = projectId
   }
 
   /**
@@ -154,20 +192,14 @@ export class CloudApi {
    */
   static async factory({
     log,
-    currentDirectory,
+    cloudDomain,
     skipLogging = false,
   }: {
     log: LogEntry
-    currentDirectory: string
+    cloudDomain: string
     skipLogging?: boolean
   }) {
     log.debug("Initializing Garden Cloud API client.")
-
-    const config = await getEnterpriseConfig(currentDirectory)
-    if (!config) {
-      log.debug("Cloud/Enterprise domain and/or project ID missing. Aborting.")
-      return null
-    }
 
     const token = await CloudApi.getClientAuthTokenFromDb(log)
     if (!token && !gardenEnv.GARDEN_AUTH_TOKEN) {
@@ -175,10 +207,10 @@ export class CloudApi {
       return null
     }
 
-    const api = new CloudApi(log, config.domain, config.projectId)
+    const api = new CloudApi(log, cloudDomain)
     const tokenIsValid = await api.checkClientAuthToken()
 
-    const distroName = getCloudDistributionName(config.domain)
+    const distroName = getCloudDistributionName(api.domain)
     const section = distroName === "Garden Enterprise" ? "garden-enterprise" : "garden-cloud"
 
     const enterpriseLog = skipLogging ? null : log.info({ section, msg: "Authorizing...", status: "active" })
@@ -217,13 +249,6 @@ export class CloudApi {
     }
 
     enterpriseLog?.setSuccess({ msg: chalk.green("Done"), append: true })
-    try {
-      const project = await api.getProject()
-      const url = new URL(`/projects/${project.id}`, api.domain)
-      enterpriseLog?.info({ symbol: "info", msg: `Visit project at ${url.href}` })
-    } catch (err) {
-      log.debug(`Getting project from API failed with error: ${err.message}`)
-    }
 
     return api
   }
@@ -336,6 +361,98 @@ export class CloudApi {
       clearInterval(this.intervalId)
       this.intervalId = null
     }
+  }
+
+  /**
+   * Verifies the projectId against Garden Cloud and assigns it
+   * to the active API instance. Returns the project metadata or throws
+   * an error if the project does not exist.
+   */
+  async verifyAndConfigureProject(projectId: string): Promise<CloudProject> {
+    let project: CloudProject | undefined
+    try {
+      this.projectId = projectId
+      project = await this.getProject()
+    } catch (err) {
+      this.projectId = undefined
+      throw err
+    }
+
+    if (!project) {
+      throw new EnterpriseApiError(`Garden Cloud has no project with ${projectId}`, {})
+    }
+
+    return project
+  }
+
+  async getProjectByName(projectName: string): Promise<CloudProject | undefined> {
+    let response: ListProjectsResponse
+
+    try {
+      response = await this.get<ListProjectsResponse>(`/projects`)
+    } catch (err) {
+      this.log.debug(`Attempt to list all projects failed with ${err}`)
+      throw err
+    }
+
+    let projects: ListProjectsResponse["data"] = response.data
+
+    projects = projects.filter((p) => p.name === projectName)
+
+    // Expect a single project, otherwise we fail with an error
+    if (projects.length > 1) {
+      throw new EnterpriseApiDuplicateProjectsError(
+        deline`Found an unexpected state with multiple projects using the same name, ${projectName}.
+        Please make sure there is only one project with the given name.
+        Projects can be deleted through the Garden Cloud UI at ${this.domain}`,
+        {}
+      )
+    }
+
+    let project: ListProjectsResponse["data"][0] | undefined = projects[0]
+
+    if (!project) {
+      return undefined
+    }
+
+    return toCloudProject(project)
+  }
+
+  async createProject(projectName: string): Promise<CloudProject> {
+    let response: CreateProjectsForRepoResponse
+
+    try {
+      const createRequest = {
+        name: projectName,
+        repositoryUrl: "",
+        relativeProjectRootPath: "",
+        importFromVcsProvider: false,
+      }
+
+      response = await this.post<CreateProjectsForRepoResponse>(`/projects/`, {
+        body: createRequest,
+      })
+    } catch (err) {
+      this.log.debug(`Create project request failed with error, ${err}`)
+      throw err
+    }
+
+    const project: CreateProjectsForRepoResponse["data"][0] = response.data[0]
+    return toCloudProject(project)
+  }
+
+  async getOrCreateProject(projectName: string): Promise<CloudProject> {
+    let project: CloudProject | undefined = await this.getProjectByName(projectName)
+
+    if (!project) {
+      project = await this.createProject(projectName)
+    }
+
+    // This is necessary to internally configure the project for this instance
+    this._project = project
+    this.projectId = project.uid
+
+    return project
   }
 
   private async refreshTokenIfExpired() {
@@ -535,7 +652,7 @@ export class CloudApi {
         environment,
         namespace,
       }
-      this.log.debug("Registering session with Garden Cloud.")
+      this.log.debug(`Registering session with Garden Cloud for ${this.projectId} in ${environment}/${namespace}.`)
       const res: RegisterSessionResponse = await this.post("sessions", {
         body,
         retry: true,
@@ -561,13 +678,25 @@ export class CloudApi {
     this.sessionRegistered = true
   }
 
-  async getProject() {
-    if (this._project) {
+  async getProject(): Promise<CloudProject | undefined> {
+    if (!this.projectId) {
+      this.log.debug(`Could not retrieve a project which has not yet been configured`)
+      return
+    }
+
+    // If we are using a new project ID, retrieve again from the API
+    // NOTE: If we wan't to use this with multiple project IDs we need
+    // a cache supporting that + check if the remote project metadata
+    // was updated.
+    if (this._project && this._project.uid === this.projectId) {
       return this._project
     }
 
     const res = await this.get<GetProjectResponse>(`/projects/uid/${this.projectId}`)
-    this._project = res.data
+    const project: GetProjectResponse["data"] = res.data
+
+    this._project = toCloudProject(project)
+
     return this._project
   }
 
@@ -576,7 +705,7 @@ export class CloudApi {
       return this._profile
     }
 
-    const res = await this.get<GetUserResponse>(`/profile`)
+    const res = await this.get<GetProfileResponse>(`/profile`)
     this._profile = res.data
     return this._profile
   }