+++ core/test/integ/src/plugins/kubernetes/helm/deployment.ts
@@ -318,7 +318,7 @@ describe("deployHelmService", () => {
   })
 
   it("should mark a chart that has been paused by Garden Cloud AEC as outdated", async () => {
-    const fakeCloudApi = new CloudApi(getLogger().placeholder(), "https://test.cloud.garden.io", "project-id")
+    const fakeCloudApi = new CloudApi(getLogger().placeholder(), "https://test.cloud.garden.io")
     const projectRoot = resolve(dataDir, "test-projects", "helm")
     const gardenWithCloudApi = await makeTestGarden(projectRoot, { cloudApi: fakeCloudApi, noCache: true })